moved input to start page
