+++ src/app/game/page.tsx
@@ -7,16 +7,11 @@ import React, { useState } from "react";
 const Game = () => {
   const [isGameOngoing, setIsGameOnGoing] = useState<boolean>(false);
 
-  const [playerName, setPlayerName] = useState<string>("")
+  
 
   const [currentPoints, setCurrentPoints] = useState(0);
         
   const [showCountdown, setShowCountdown] = useState<boolean>(false);
-
-
-  const handleOnChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setPlayerName(event.target.value);
-  }; 
 
   //funktion för att öka poängen, kan användas i spellogiken när en mullvad träffas.
   const addPoint = () => {
@@ -36,11 +31,7 @@ const Game = () => {
 
   return (
     <div>
-      <Input 
-      value={playerName} 
-      onChange={handleOnChange}
-      disabled={isGameOngoing}
-      />
+      
       <StartButton
         btnText="Start a New Game"
         onClick={handleNewGame}
